Organised java files
