+++ app/build.gradle
@@ -32,8 +32,8 @@ android {
     productFlavors {
         PassIssuer {
             dimension = 'version'
-            versionCode = 3
-            versionName = '0.0.3'
+            versionCode = 4
+            versionName = '0.0.4'
             targetSdkVersion 29
             minSdkVersion 21
             applicationIdSuffix = '.passissuer'
@@ -41,8 +41,8 @@ android {
         }
         PassScanner {
             dimension = 'version'
-            versionCode = 3
-            versionName = '0.0.3'
+            versionCode = 4
+            versionName = '0.0.4'
             targetSdkVersion 29
             minSdkVersion 21
             applicationIdSuffix = '.passscanner'
@@ -50,8 +50,8 @@ android {
         }
         PassUser {
             dimension = 'version'
-            versionCode = 3
-            versionName = '0.0.3'
+            versionCode = 4
+            versionName = '0.0.4'
             targetSdkVersion 29
             minSdkVersion 21
             applicationIdSuffix = '.passuser'
