+++ app/src/main/java/org/covid19india/android/safepassageindia/model/User.java
@@ -1,4 +1,4 @@
-package org.covid19india.android.safepassageindia;
+package org.covid19india.android.safepassageindia.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
